+++ src/components/myTeam/MyTeam.vue
@@ -3,8 +3,7 @@
       <h1>소속팀 목록</h1>
       <!-- Error :  Avoid using non-primitive value as key, use string/number value instead. -->
       <!-- 해결법 : https://crispypotato.tistory.com/33 -->
-      <v-card class="mt-10" v-for="(team,index) in teamList" v-bind:key="index">    
-        <MyTeamPage :teamSeq="team.teamSeq"></MyTeamPage>    
+      <v-card class="mt-10">    
         <v-row class="mb-5">
           <v-col align-self="center" md="3">
             <v-img
@@ -44,42 +43,17 @@
           <v-btn class="float-right" to="myTeam">상세보기</v-btn>
         </v-col>                
       </v-card>      
-      <!-- <v-btn v-on:click="load">버튼</v-btn> -->
   </v-container>
 </template>
 
 <script>
-// Vue lifeCycle 에 관하여
-// https://wormwlrm.github.io/2018/12/29/Understanding-Vue-Lifecycle-hooks.html
-import MyTeamApi from '@/api/MyTeamAPI'
-import MyTeamPage from '@/views/myTeam/MyTeamPage.vue';
-
 export default {
-  data: ()=>{
-    return {
-      teamList: []
-    }
-  },
-  components:{
-    MyTeamPage,
-  },
-  methods:{
-    async load(){
-      // 비동기적인 console.log 처리로 인해 발생하는 현상
-      // https://kkangdda.tistory.com/81
-      try{
-        const list = await MyTeamApi.searchTeams();
-        list.data.forEach(element => {
-          this.teamList.push(element);
-        });
-      }catch(error){
-        console.log(error);
-      }      
-    }
-  },
-  mounted () {
-    this.load();
-  }
+    data:()=>{
+        return {
+
+        }
+    },
+    props:['teamInfo']
 }
 </script>
